Cleanup ingester indexing logic, better comments
